# bandcoord/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'BandCoord API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://www.iestrassierra.net/alumnado/curso2425/DAW/daw2425a16/laravel'
...
paths:
...
  '/api/usuarios/{id}/approve':
    patch:
      summary: 'Aprobar usuario'
      operationId: aprobarUsuario
      description: "Cambia el estado de un usuario a 'activo'."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Usuario aprobado correctamente.'
                  data:
                    id: 1
                    estado: activo
                properties:
                  message:
                    type: string
                    example: 'Usuario aprobado correctamente.'
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      estado:
                        type: string
                        example: activo
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Usuario no encontrado.'
                properties:
                  error:
                    type: string
                    example: 'Usuario no encontrado.'
      tags:
        - 'Gestión de Usuarios'
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'El ID del usuario.'
        example: '1'
        required: true
        schema:
          type: string
...